libcloud-compute/pyproject.toml
[tool.poetry]
name = "libcloud-compute"
version = "0.1.0"
description = ""
authors = ["Nathan <thetruecoderman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
apache-libcloud = "^3.6.0"
requests-debugger = "^0.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
